Testing in Matlab Fixed issue with parameter not updating based on specified class.

+++ titans.nam/src/main/java/titans/nam/grib/GribFile.java
@@ -87,6 +87,9 @@ public class GribFile {
    */
   public boolean isLocked() {
     try {
+      if (this.notExists()) {
+        return false;
+      }
       FileOutputStream outputStream = new FileOutputStream(this.grib);
       FileChannel channel = outputStream.getChannel();
       FileLock lock;
